extract marker list into seperate component
