+++ src/screens/DashboardScreen.js
@@ -4,6 +4,7 @@ import {
   Text,
   Image,
   Button,
+  FlatList,
   ScrollView /* AsyncStorage */
 } from 'react-native';
 import { connect } from 'react-redux';
@@ -16,15 +17,11 @@ import {
   deleteMarkerDashboard
 } from '../actions/markerActions';
 import { getCurrentUser } from '../actions/userActions';
+import MarkerList from '../common/MarkerList';
 
 import * as styles from '../styles/dashboardStyles';
 
 class DashboardScreen extends Component {
-  componentDidMount() {
-    this.props.dispatch(getCurrentUser());
-    this.props.dispatch(getMarkersDashboard());
-  }
-
   constructor(props) {
     super(props);
     this.state = {
@@ -32,38 +29,7 @@ class DashboardScreen extends Component {
     };
   }
 
-  onClick(markerId) {
-    this.props.dispatch(deleteMarkerDashboard({ markerId }));
-  }
-
   render() {
-    const reports = this.props.markersFromServer;
-    let username;
-
-    let userReports = reports.map(report => (
-      <View style={styles.reportCard} key={report._id}>
-        {/* {this.props.loggedIn ? '' : (<Redirect to='/map' />)} */}
-        <Text style={styles.incidentType}>{report.incidentType}</Text>
-        <Image
-          style={styles.reportIcon}
-          alt={`Report icon for ${report.incidentType}`}
-          source={{ uri: 'report.icon' }}
-        />
-        <Text style={styles.incidentDate}>Date: {report.date}</Text>
-        <Text style={styles.incidentDescriptionTitle}>Description:</Text>
-        <Text style={styles.incidentDescription}>{report.description}</Text>
-        <Button
-          onPress={() => this.onClick(report._id)}
-          title="Delete"
-          id={report._id}
-          style={styles.deleteIncident}
-          accessibilityLabel="Delete this incident report"
-        >
-          Delete
-        </Button>
-      </View>
-    ));
-
     return (
       <View style={styles.view}>
         <Navbar
@@ -72,24 +38,13 @@ class DashboardScreen extends Component {
           back={() => this.props.navigation.goBack()}
         />
         <UserProfile navigation={this.props.navigation} />
-        <ScrollView>
-          <View style={styles.userReports}>{userReports}</View>
-        </ScrollView>
+        <MarkerList />
       </View>
     );
   }
 }
 
-export const mapStateToProps = state => {
-  console.log('DASHBOARD SCREEN STATE:', state);
-  return {
-    markersFromServer: state.markers.allMarkers ? state.markers.allMarkers : [],
-    loggedIn: state.auth.currentUser !== null,
-    currentUser: state.auth.currentUser ? state.auth.currentUser : ''
-  };
-};
-
-export default connect(mapStateToProps)(DashboardScreen);
+export default DashboardScreen;
 
 /*
 Resources:
